Bugfixing in encrypting existing keys
