+++ lib/askuser.js
@@ -1,5 +1,8 @@
+require('dotenv').config();
 const chalk = require('chalk');
 const inquirer = require('inquirer');
+const { find } = require('./database');
+const { createPasswordList } = require('./createlists');
 
 async function askUser(question) {
     const { answer } = await inquirer.prompt([
@@ -60,6 +63,20 @@ async function askPassword () {
     return pwdInput
 }
 
+async function choosePassword(categories) {
+    const choosenCategory = await showOptions(
+        categories,
+        `Choose a category from below for continuing`
+    );
+
+    const documents = await find(process.env.DB_COLLECTION, { category: choosenCategory });
+    const choices = await createPasswordList(documents);
+    const passwordID = await showOptions(choices, 'Choose a password.');
+
+    return passwordID;
+}
+
+exports.choosePassword = choosePassword;
 exports.askUser = askUser;
 exports.getUserChoice = getUserChoice;
 exports.askPassword = askPassword;
