+++ lib/createlists.js
@@ -12,9 +12,9 @@ function createCategoryList(collection) {
 
 function createPasswordList(documents) {
     const choices = documents.map((document) => {
-        return { 
+        return {
             name: document.name,
-            value: document._id
+            value: document._id,
         };
     });
     return choices;
