+++ lib/database.js
@@ -40,9 +40,9 @@ async function deleteDocument(name, id) {
     }
 }
 
-async function changeDocument(name, id, newValue) {
+async function changeDocument(name, passwordID, newValue) {
     try {
-        await db.collection(name).updateOne({ _id: id }, { $set: { value: newValue } });
+        await db.collection(name).updateOne({ _id: passwordID }, { $set: { value: newValue } });
     } catch (err) {
         console.log(err);
         return;
